Add checking if object exists to Update/Delete action

+++ car-themed-app-solution/car-themed-app/Handlers/Orders/DeleteOrderByIdHandler.cs
@@ -1,12 +1,15 @@
 ﻿using car_themed_app.Commands.Orders;
+using car_themed_app_Repository;
 using car_themed_app_Repository.Interfaces;
+using car_themed_app_Repository.Models;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
+using static car_themed_app_Repository.GenericResultMethods;
 
 namespace car_themed_app.Handlers.Orders
 {
-    public class DeleteOrderByIdHandler : IRequestHandler<DeleteOrderByIdCommand>
+    public class DeleteOrderByIdHandler : IRequestHandler<DeleteOrderByIdCommand, Result<string>>
     {
         private readonly IOrdersRepository _ordersRepository;
 
@@ -15,10 +18,15 @@ namespace car_themed_app.Handlers.Orders
             _ordersRepository = ordersRepository;
         }
 
-        public Task<Unit> Handle(DeleteOrderByIdCommand request, CancellationToken cancellationToken)
+        public async Task<Result<string>> Handle(DeleteOrderByIdCommand request, CancellationToken cancellationToken)
         {
-            _ordersRepository.DeleteOrder(request.Id);
-            return Task.FromResult(Unit.Value);
+            Order order = await _ordersRepository.GetOrderAsync(request.Id);
+            if(order != null)
+            {
+                _ordersRepository.DeleteOrder(request.Id);
+                return Ok<string>();
+            }
+            return ErrorMessage<string>($"Order {request.Id} does not exist.");
         }
     }
 }
